requirements.txt
Django
django-filter
djangorestframework
django-decouple
django-ckeditor
drf-spectacular
django-widget-tweaks
django-redis
pillow
python-decouple
django-cleanup
django-cors-headers
requests
psycopg2-binary
msal
djangorestframework-simplejwt
django-allauth
psycopg2-binary
gunicorn
django-crontab
django-simple-history
pandas
minio
openpyxl
celery[redis]
redis
